Add files via upload
This project is a simple and intuitive To-Do List web application built using HTML, CSS, and JavaScript. The app allows users to add and delete tasks, as well as mark them as completed. It features a clean and responsive design, making it accessible across various devices. This project highlights my skills in front-end development and demonstrates my ability to create interactive and user-friendly web applications.